Abstract crypto code via features and capabilities
If a crypto provider is not configured via feature (currently only
`crypto-ring`), all crypto related states in the protocol will be skipped by the
requester. If crypto capabilities are enabled in the config.toml file, but a
provider feature is not given, build.rs will panic.

An attempt was made to minimize the amount of conditional compilation. Almost
all of the conditions exist in the crypto modules themselves with default
implementations that will panic if run by higher level code. This is safe for
two reasons:

 1. States that use crypto will be skipped if the capabilities are not enabled.
 2. Build.rs will panic if a crypto provider is not enabled, but capabilities
 that use crypto are enabled.

The `Nonce` code was also moved out of `src/msgs/common.rs` and into
`src/crypto/common.rs` to keep the message code mostly free of conditional
compilation.

Also, reliance on ring was removed for certificate chain parsing of DER
encoded chains. This was easy as it only required adding consts for two DER tag
bytes.

Lastly, a crypto provider trait was added, but is not yet used. We plan to
transition the existing crypto traits to the de-facto standard rust-crypto
traits in a follow up request, and we will almost certainly use the provider at
that point. That will remove the need for use of silly type aliases like
`DigestImpl` that are behind features.
