+++ Cargo.toml
@@ -2,20 +2,25 @@
 name = "spdm"
 version = "0.1.0"
 edition = "2018"
-authors = ["Andrew J. Stone <andrew@oxidecomputer.com>"]
 license = "MPL 2.0"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[features]
+crypto-ring = ["ring", "webpki", "rand"]
 
 [dependencies]
 bitflags = "1.3"
-ring = "0.16.20"
-webpki = "0.22.0"
+webpki = { version = "0.22.0", optional = true}
+
+[dependencies.ring]
+version = "0.16.20"
+default-features = false
+optional = true
 
 [dependencies.rand]
 version = "0.8"
 default-features = false
 features = [ "getrandom" ]
+optional = true
 
 [dev-dependencies]
 test-utils = { path = "test-utils" }
