+++ src/crypto/ring/mod.rs
@@ -1,0 +1,9 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! Crypto backed by <https://github.com/briansmith/ring>
+
+pub mod digest;
+pub mod pki;
+pub mod signing;
