+++ src/msgs/common.rs
@@ -8,7 +8,6 @@ use super::{BufferFullError, ReadError, Reader, Writer};
 use crate::config;
 
 use core::convert::{From, TryFrom, TryInto};
-use rand::{rngs::OsRng, RngCore};
 
 #[derive(Debug, PartialEq)]
 pub enum ParseOpaqueDataError {
@@ -494,36 +493,6 @@ impl PartialEq for SignatureBuf {
 
 impl Eq for SignatureBuf {}
 
-/// A unique random value used for cryptographic purposes
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Nonce([u8; 32]);
-
-impl Nonce {
-    pub fn new() -> Nonce {
-        let mut nonce = [0u8; 32];
-        OsRng.fill_bytes(&mut nonce);
-        Nonce(nonce)
-    }
-
-    pub fn read(r: &mut Reader) -> Result<Nonce, ReadError> {
-        let mut nonce = [0u8; 32];
-        r.get_slice(32, &mut nonce)?;
-        Ok(Nonce(nonce))
-    }
-}
-
-impl AsRef<[u8]> for Nonce {
-    fn as_ref(&self) -> &[u8] {
-        &self.0
-    }
-}
-
-impl AsMut<[u8]> for Nonce {
-    fn as_mut(&mut self) -> &mut [u8] {
-        &mut self.0
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -539,12 +508,6 @@ mod tests {
     impl SignatureBuf {
         pub fn new_with_magic(size: SignatureSize, magic: u8) -> SignatureBuf {
             SignatureBuf { size, buf: [magic; config::MAX_SIGNATURE_SIZE] }
-        }
-    }
-
-    impl Nonce {
-        pub fn new_with_magic(magic: u8) -> Nonce {
-            Nonce([magic; 32])
         }
     }
