+++ src/responder/no_crypto_states.rs
@@ -1,0 +1,7 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! This module contains the minimal implementation of states that are not in
+//! use when the "crypto" feature  is not enabled. The purpose of this is to
+//! minimize the amount of conditional compilation required in the rest of the code.
